Premier envoie du projet

+++ app/Http/Controllers/CategorieController.php
@@ -2,8 +2,8 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
 use App\Models\Categorie;
+use Illuminate\Http\Request;
 
 class CategorieController extends Controller
 {
@@ -13,10 +13,9 @@ class CategorieController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function index()
-    {
+    {     
         $category = Categorie::all();
-    return view('categories.index', ['category' => $category]);
-   
+        return view('categories.index', ['category' => $category]);
     }
 
     /**
@@ -37,10 +36,26 @@ class CategorieController extends Controller
      */
     public function store(Request $request)
     {
-        $categorie = new Categorie;
-        $categorie->nom = $request->input('nom');
-        $categorie->save();
-        
+
+        $user = $request->user();
+
+        $type_role_user = $user->role->where('nom', 'user');
+        $type_role_admin = $user->role->where('nom', 'admin');
+        $type_role_moderateur = $user->role->where('nom', 'moderateur');
+
+        $isAdmin = $type_role_admin->count() > 0;
+        $isUser = $type_role_user->count() > 0;
+        $isModerateur = $type_role_moderateur->count() > 0;
+
+
+        if ($isAdmin) {
+            $categorie = new Categorie;
+            $categorie->nom = $request->input('nom');
+            $categorie->save();
+        } else {
+            return redirect()->route('categorie.index')->with('danger', 'Seul un admin peut créer de catégories.');;
+        }
+
         return redirect()->route('categorie.index')->with('success', 'Catégorie creé avec succès.');
     }
 
@@ -64,7 +79,7 @@ class CategorieController extends Controller
      */
     public function edit($id)
     {
-        $category =Categorie::findOrFail($id);
+        $category = Categorie::findOrFail($id);
 
         return view('categories.edit', ['category' => $category]);
     }
@@ -78,19 +93,29 @@ class CategorieController extends Controller
      */
     public function update(Request $request, $id)
     {
-    //     // Valide les données envoyées par le formulaire
-    //     $request->validate([
-    //     'nom' => 'required|string|max:255'
-    // ]);
+        $user = $request->user();
 
-    // Met à jour les données de la catégorie
-    $category = Categorie::findOrFail($id);
-    $category->nom = $request->input('nom');
-    $category->save();
+        $type_role_user = $user->role->where('nom', 'user');
+        $type_role_admin = $user->role->where('nom', 'admin');
+        $type_role_moderateur = $user->role->where('nom', 'moderateur');
 
-    // Redirige vers la vue index avec un message de succès
-    return redirect()->route('categorie.index')
-                     ->with('success', 'Catégorie mise à jour avec succès.');
+        $isAdmin = $type_role_admin->count() > 0;
+        $isUser = $type_role_user->count() > 0;
+        $isModerateur = $type_role_moderateur->count() > 0;
+
+        if ($isAdmin) {
+            // Met à jour les données de la catégorie
+            $category = Categorie::findOrFail($id);
+            $category->nom = $request->input('nom');
+            $category->save();
+        } else {
+            return redirect()->route('categorie.index')
+                ->with('danger', 'Seuls les admins peuvent mettre à jour.');
+        }
+
+        // Redirige vers la vue index avec un message de succès
+        return redirect()->route('categorie.index')
+            ->with('success', 'Catégorie mise à jour avec succès.');
     }
 
     /**
@@ -99,12 +124,27 @@ class CategorieController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy($id, Request $request)
     {
-        $category = Categorie::findOrFail($id);
-        $category->delete();
-       
-    
+        $user = $request->user();
+
+        $type_role_user = $user->role->where('nom', 'user');
+        $type_role_admin = $user->role->where('nom', 'admin');
+        $type_role_moderateur = $user->role->where('nom', 'moderateur');
+
+        $isAdmin = $type_role_admin->count() > 0;
+        $isUser = $type_role_user->count() > 0;
+        $isModerateur = $type_role_moderateur->count() > 0;
+
+        if ($isAdmin) {
+            // Met à jour les données de la catégorie
+            $category = Categorie::findOrFail($id);
+            $category->delete();
+        } else {
+            return redirect()->route('categorie.index')
+                ->with('danger', 'Seuls les admins peuvent supprimer des catégories.');
+        }
+
         return redirect()->route('categorie.index')->with('success', 'Category deleted!');
     }
 }
